Remove unecessary sqlite references and other code following PR
