+++ .travis.yml
@@ -29,8 +29,6 @@ deploy:
   environment:
     variables:
       LOG_LEVEL: debug
-      SQLITE_FILE: checkInCards.sql
-      SQLITE_BUCKET: check-in-card-data-qa
       DB_HOST: 10.146.200.10
       DB_PORT: '1032'
       DB_PSWD: AQECAHh7ea2tyZ6phZgT4B9BDKwguhlFtRC6hgt+7HbmeFsrsgAAAGwwagYJKoZIhvcNAQcGoF0wWwIBADBWBgkqhkiG9w0BBwEwHgYJYIZIAWUDBAEuMBEEDDnjoRlSmwePv1QxSwIBEIAplnBJKYjr0qYPnQ3GIZ3yDzbf3i0Aa5q+dLB07MRFxbF6Rok1MKeCwZ8=
@@ -68,8 +66,6 @@ deploy:
   environment:
     variables:
       LOG_LEVEL: debug
-      SQLITE_FILE: checkInCards.sql
-      SQLITE_BUCKET: check-in-card-data-production
       LOG_LEVEL: info
       DB_HOST: 10.146.200.8
       DB_PORT: '1032'
