+++ sam.dev.yml
@@ -17,8 +17,3 @@ Globals:
 Resources:
   SierraCheckInCardPoller-dev:
     Type: AWS::Serverless::Function
-    Properties:
-      Environment:
-        Variables:
-            SQLITE_FILE: 'checkInCards.sql'
-            SQLITE_BUCKET: 'check-in-card-data'
